requirements.txt
Django==1.10.4
django-filter==1.0.1
django-webpack-loader==0.4.1
djangorestframework==3.5.3
Markdown==2.6.7
Pillow==3.4.2
